move fb functions into crud files
