+++ src/firebase/picks/get.ts
@@ -1,0 +1,5 @@
+/**
+ * these are contained in a users profile
+ * but may add a function that returns only specified week
+ */
+export {}
