+++ src/firebase/picks/update.ts
@@ -1,0 +1,4 @@
+/**
+ * see ./create.ts
+ */
+export {};
