+++ src/firebase/slate/update.ts
@@ -1,0 +1,1 @@
+export {};
